move /me endpoints to unique /auth/me endpoint
