+++ app/utils/variables.py
@@ -2,13 +2,12 @@ DEFAULT_APP_NAME = "Albert API"
 DEFAULT_TIMEOUT = 300
 
 ENDPOINT__ADMIN_ROLES = "/admin/roles"
-ENDPOINT__ADMIN_ROLES_ME = "/admin/roles/me"
 ENDPOINT__ADMIN_TOKENS = "/admin/tokens"
 ENDPOINT__ADMIN_USERS = "/admin/users"
-ENDPOINT__ADMIN_USERS_ME = "/admin/users/me"
 ENDPOINT__AGENTS_COMPLETIONS = "/agents/completions"
 ENDPOINT__AGENTS_TOOLS = "/agents/tools"
 ENDPOINT__AUDIO_TRANSCRIPTIONS = "/audio/transcriptions"
+ENDPOINT__AUTH_ME = "/auth/me"
 ENDPOINT__CHAT_COMPLETIONS = "/chat/completions"
 ENDPOINT__CHUNKS = "/chunks"
 ENDPOINT__COLLECTIONS = "/collections"
